+++ firestore-bigquery-export/functions/jest.config.js
@@ -16,7 +16,15 @@ module.exports = {
   moduleNameMapper: {
     "firebase-admin/eventarc":
       "<rootDir>/node_modules/firebase-admin/lib/eventarc",
+    "firebase-admin/firestore":
+      "<rootDir>/node_modules/firebase-admin/lib/firestore",
     "firebase-functions/v2": "<rootDir>/node_modules/firebase-functions/lib/v2",
+    "firebase-admin/auth": "<rootDir>/node_modules/firebase-admin/lib/auth",
+    "firebase-admin/app": "<rootDir>/node_modules/firebase-admin/lib/app",
+    "firebase-admin/database":
+      "<rootDir>/node_modules/firebase-admin/lib/database",
+    "firebase-admin/firestore":
+      "<rootDir>/node_modules/firebase-admin/lib/firestore",
     "firebase-admin/functions":
       "<rootDir>/node_modules/firebase-admin/lib/functions",
     "firebase-admin/extensions":