+++ firestore-bigquery-export/functions/src/index.ts
@@ -16,18 +16,19 @@
 
 import config from "./config";
 import * as functions from "firebase-functions";
+import * as admin from "firebase-admin";
+import { getExtensions } from "firebase-admin/extensions";
+import { getFunctions } from "firebase-admin/functions";
 import {
   ChangeType,
   FirestoreBigQueryEventHistoryTracker,
   FirestoreEventHistoryTracker,
 } from "@firebaseextensions/firestore-bigquery-change-tracker";
 
-import * as admin from "firebase-admin";
-import { getFunctions } from "firebase-admin/functions";
-import { getExtensions } from "firebase-admin/extensions";
 import { getEventarc } from "firebase-admin/eventarc";
 import * as logs from "./logs";
-import { getChangeType, getDocumentId } from "./util";
+import * as events from "./events";
+import { getChangeType, getDocumentId, resolveWildcardIds } from "./util";
 
 const eventTracker: FirestoreEventHistoryTracker =
   new FirestoreBigQueryEventHistoryTracker({
@@ -45,6 +46,7 @@ const eventTracker: FirestoreEventHistoryTracker =
     bqProjectId: config.bqProjectId,
     useNewSnapshotQuerySyntax: config.useNewSnapshotQuerySyntax,
     skipInit: true,
+    kmsKeyName: config.kmsKeyName,
   });
 
 logs.init();
@@ -54,11 +56,7 @@ if (admin.apps.length === 0) {
   admin.initializeApp();
 }
 
-const eventChannel =
-  process.env.EVENTARC_CHANNEL &&
-  getEventarc().channel(process.env.EVENTARC_CHANNEL, {
-    allowedEventTypes: process.env.EXT_SELECTED_EVENTS,
-  });
+events.setupEventChannel();
 
 export const syncBigQuery = functions.tasks
   .taskQueue({
@@ -68,28 +66,40 @@ export const syncBigQuery = functions.tasks
     },
     rateLimits: {
       maxConcurrentDispatches: 1000,
-      maxDispatchesPerSecond: 500,
+      maxDispatchesPerSecond: config.maxDispatchesPerSecond,
     },
   })
   .onDispatch(
     async ({ context, changeType, documentId, data, oldData }, ctx) => {
-      await eventTracker.record([
-        {
-          timestamp: context.timestamp, // This is a Cloud Firestore commit timestamp with microsecond precision.
-          operation: changeType,
-          documentName: context.resource.name,
-          documentId: documentId,
-          pathParams: config.wildcardIds ? context.params : null,
-          eventId: context.eventId,
-          data,
-          oldData,
+      const update = {
+        timestamp: context.timestamp, // This is a Cloud Firestore commit timestamp with microsecond precision.
+        operation: changeType,
+        documentName: context.resource.name,
+        documentId: documentId,
+        pathParams: config.wildcardIds ? context.params : null,
+        eventId: context.eventId,
+        data,
+        oldData,
+      };
+
+      /** Record the chnages in the change tracker */
+      await eventTracker.record([{ ...update }]);
+
+      /** Send an event Arc update , if configured */
+      await events.recordSuccessEvent({
+        subject: documentId,
+        data: {
+          ...update,
         },
-      ]);
+      });
+
+      logs.complete();
     }
   );
 
-export const fsexportbigquery = functions.firestore
-  .document(config.collectionPath)
+export const fsexportbigquery = functions
+  .runWith({ failurePolicy: true })
+  .firestore.document(config.collectionPath)
   .onWrite(async (change, context) => {
     logs.start();
     try {
@@ -102,37 +112,49 @@ export const fsexportbigquery = functions.firestore
       const data = isDeleted ? undefined : change.after.data();
       const oldData = isCreated ? undefined : change.before.data();
 
-      if (eventChannel) {
-        await eventChannel.publish({
-          type: `firebase.extensions.big-query-export.v1.sync.start`,
-          data: {
-            documentId,
-            changeType,
-            before: {
-              data: change.before.data(),
-            },
-            after: {
-              data: change.after.data(),
-            },
-            context: context.resource,
-          },
-        });
-      }
+      await events.recordStartEvent({
+        documentId,
+        changeType,
+        before: {
+          data: change.before.data(),
+        },
+        after: {
+          data: change.after.data(),
+        },
+        context: context.resource,
+      });
 
       const queue = getFunctions().taskQueue(
         `locations/${config.location}/functions/syncBigQuery`,
         config.instanceId
       );
 
+      /**
+       * enqueue data cannot currently handle documentdata
+       * Serialize early before queueing in clopud task
+       * Cloud tasks currently have a limit of 1mb, this also ensures payloads are kept to a minimum
+       */
+      const seializedData = eventTracker.serializeData(data);
+      const serializedOldData = eventTracker.serializeData(oldData);
+
       await queue.enqueue({
         context,
         changeType,
         documentId,
-        data,
-        oldData,
+        data: seializedData,
+        oldData: serializedOldData,
       });
     } catch (err) {
+      await events.recordErrorEvent(err as Error);
       logs.error(err);
+      const eventAgeMs = Date.now() - Date.parse(context.timestamp);
+      const eventMaxAgeMs = 10000;
+
+      if (eventAgeMs > eventMaxAgeMs) {
+        return;
+      }
+
+      throw err;
     }
 
     logs.complete();
@@ -141,13 +163,110 @@ export const fsexportbigquery = functions.firestore
 export const setupBigQuerySync = functions.tasks
   .taskQueue()
   .onDispatch(async () => {
-    /** Get extensions runtime */
+    /** Setup runtime environment */
     const runtime = getExtensions().runtime();
 
     /** Init the BigQuery sync */
-
     await eventTracker.initialize();
 
-    /** Set status to complete */
-    await runtime.setProcessingState("PROCESSING_COMPLETE", "Sync complete");
+    await runtime.setProcessingState(
+      "PROCESSING_COMPLETE",
+      "Sync setup completed"
+    );
+  });
+
+export const initBigQuerySync = functions.tasks
+  .taskQueue()
+  .onDispatch(async () => {
+    /** Setup runtime environment */
+    const runtime = getExtensions().runtime();
+
+    /** Init the BigQuery sync */
+    await eventTracker.initialize();
+
+    /** Run Backfill */
+    if (config.doBackfill) {
+      await getFunctions()
+        .taskQueue(
+          `locations/${config.location}/functions/fsimportexistingdocs`,
+          config.instanceId
+        )
+        .enqueue({ offset: 0, docsCount: 0 });
+      return;
+    }
+
+    await runtime.setProcessingState(
+      "PROCESSING_COMPLETE",
+      "Sync setup completed"
+    );
+    return;
+  });
+
+exports.fsimportexistingdocs = functions.tasks
+  .taskQueue()
+  .onDispatch(async (data, context) => {
+    const runtime = getExtensions().runtime();
+    if (!config.doBackfill || !config.importCollectionPath) {
+      await runtime.setProcessingState(
+        "PROCESSING_COMPLETE",
+        "Completed. No existing documents imported into BigQuery."
+      );
+      return;
+    }
+
+    const offset = (data["offset"] as number) ?? 0;
+    const docsCount = (data["docsCount"] as number) ?? 0;
+
+    const query = config.useCollectionGroupQuery
+      ? admin
+          .firestore()
+          .collectionGroup(
+            config.importCollectionPath.split("/")[
+              config.importCollectionPath.split("/").length - 1
+            ]
+          )
+      : admin.firestore().collection(config.importCollectionPath);
+
+    const snapshot = await query
+      .offset(offset)
+      .limit(config.docsPerBackfill)
+      .get();
+
+    const rows = snapshot.docs.map((d) => {
+      return {
+        timestamp: new Date().toISOString(),
+        operation: ChangeType.IMPORT,
+        documentName: `projects/${config.bqProjectId}/databases/(default)/documents/${d.ref.path}`,
+        documentId: d.id,
+        eventId: "",
+        pathParams: resolveWildcardIds(config.importCollectionPath, d.ref.path),
+        data: eventTracker.serializeData(d.data()),
+      };
+    });
+    try {
+      await eventTracker.record(rows);
+    } catch (err: any) {
+      /** If configured, event tracker wil handle failed rows in a backup collection  */
+      functions.logger.log(err);
+    }
+    if (rows.length == config.docsPerBackfill) {
+      // There are more documents to import - enqueue another task to continue the backfill.
+      const queue = getFunctions().taskQueue(
+        "fsimportexistingdocs",
+        process.env.EXT_INSTANCE_ID
+      );
+      await queue.enqueue({
+        offset: offset + config.docsPerBackfill,
+        docsCount: docsCount + rows.length,
+      });
+    } else {
+      // We are finished, set the processing state to report back how many docs were imported.
+      runtime.setProcessingState(
+        "PROCESSING_COMPLETE",
+        `Successfully imported ${
+          docsCount + rows.length
+        } documents into BigQuery`
+      );
+    }
+    await events.recordCompletionEvent({ context });
   });